Fix group identifier retrieval
Instead of adding the identifier, respectively the name of the group
into the list of validated groups I have added the group object itself
which lead to erroneously behavior. This commit fixes it and takes the
API version into account.

Besides, a user has to be defined when validating the groups. In this
case the default user is harvest.

Signed-off-by: Konrad Reiche <konrad.reiche@fokus.fraunhofer.de>