Make remote group import configuration sensitive
In order to make the remote group import unobtrusive, the decision for
copying remote groups will be delegated to the harvester configuration.

The default case is, that remote groups will not be imported. If remote
groups should be imported which are present in local system the
configuration for "remote_groups" has to be set to "only_local".

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>