Add option to create non-existent remote groups
To complete this feature I have added another option for the remote
groups configuration field: create. In this case, the CkanCient is used
to retrieve the remote group details from the source CKAN.

Some fields have to be dropped, for instance a group's display name is
not intended to be set on creation time. These fields are dropped before
the dictionary containing the group details is passed to the CKAN action
logic.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>