Handle group not found error properly
The problem with the previous implementation was, that if the harvest
job is configured to 'only_local' and the group is not found, then the
finally block will still try to add the group to the list of validated
groups.

Besides that this will lead to a referenced before assigned error as
well, due to the nature of try-except-finally.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>