[#85] Mock web request with site_user
As described on ckan/ckan#1500, auth checking at the logic layer relies
on a `c` object being present. This is only the case in the context of a
web request. When running the harvesters via the command line this
object is not there. This means, that we can not provide a user to the
datapusher actions, as these are decoupled from our call (they use a
separate plugin interface and use a different context object).
This patch creates a `c` context object when running a harvester command
and adds the site user to it, so all furhter auth checks will be
successful.